Fixed a compilation warning in netschedule_api_submitter.cpp.
git-svn-id: https://anonsvn.ncbi.nlm.nih.gov/repos/v1/trunk/c++@49890 78c7ea69-d796-4a43-9a09-de51944f1b03

+++ src/connect/services/netschedule_api_submitter.cpp
@@ -180,14 +180,14 @@ void CNetScheduleSubmitter::SubmitJobBatch(vector<CNetScheduleJob>& jobs)
         // Batch size should be reasonable not to trigger network timeout
         const unsigned kMax_Batch = 10000;
 
-        unsigned batch_size = jobs.size() - i;
+        size_t batch_size = jobs.size() - i;
         if (batch_size > kMax_Batch) {
             batch_size = kMax_Batch;
         }
 
         cmd.erase();
         cmd = "BTCH ";
-        cmd.append(NStr::UIntToString(batch_size));
+        cmd.append(NStr::UIntToString((unsigned) batch_size));
 
         exec_result.conn->WriteLine(cmd);
